nlp-server/requirements.txt
Flask==3.0.2
scikit-learn==1.4.0
numpy==1.26.4
konlpy==0.6.0
spacy==3.7.4
python-dotenv==1.0.1
flask-cors==4.0.0 